Started work on the website
Added a layout file so that I can have components like the nav bar and the footer on every page.

Made the theme toggler component to toggle the team and created a store to store the user's theme preferences.

Added the boilerplate for the theme colours in the main app HTML file.