Renamed the guides.ts file to lists.ts
This is to better represent what the guides.ts file was supposed to be, just a list of links and files to return to the user to use.

Also added a lists folder to contain all the data in the new lists.ts file so that the data is accessible from the website.