Moved the data in the lists.ts to the data folder
Also created the required enums and the types for the data, as well as the function to return the message and the files.

Edited all the function calls in the bot's index.ts file to work with the new function.

The rest of the old functions in the lists.ts are kept for now just in case the data has any errors when being migrated. After testing and checking, the rest of the functions should be removed.