Moved the hamburger menu to the nav bar
Extracted the sidebar out of the hamburger menu and placed it in its own component.

Styled the main icon so that it will follow the size of the parent container.

Made the hover colour the normal colour and vice versa and also lightened up the dark mode hover colour in the theme toggler.

Made a table of contents element that will generate the table of contents from the page.

Added a types file that contains all the shared types used on the webpage.