Completed the side bar
Completed the table of contents as well as the collapsible menu component.

Also changed the "hyper" in the hvper sport discount catalogue to "hvper"

Added a media query to detect if the user wants reduced motion, and remove all the animations on the nav bar when it's in effect.

Positioned the nav bar where it should be, fixed at the top of the page.

Also gave the nav bar some padding.

Decreased the size of the text of the website name.

Added a hover colour to the hamburger icon.

Made the width of the sidebar the width of the content inside it.

Added some margin and padding to the sidebar.

Made all the list elements in the sidebar look just like regular text.

Changed the theme toggler icon colour to the icon colour defined in the main app.html file.

Added additional padding to the content wrapper so that there is space for the nav bar.

Added the --nav-bar-height custom CSS property so that the page content can be placed below the nav bar.

Added an --icon-hover-colour custom CSS property for the colour of the icon when it's hovered.

Changed the icon colour in dark mode to a darker shade so it's not so blinding when the opacity is 1 and changed the icon opacity to 1 so the icons aren't transparent.

Removed the types.ts file as it is no longer needed.