Made the source command easier to maintain
Used a list of hyperlinks and joined them with new lines for the source command instead of writing them into the message directly.

Removed the styles for the links in the sidebar.

The help for the /where_to_rent command now says it’s for Singapore only.

Removed the opinion in the /qr_code command.

Added a CollapsibleMenu component for the sidebar.