Updated the skate recommendations
All the prices in the skate recommendations have been updated and the recommendation for the Flying Eagle F5S has been changed to the Flying Eagle F6S due to the former increasing in price.

Also updated the review of the Nexus Phantom to state that the wheels have huge cores and to recommend that the buyer get additional wheels from other brands.

Changed the name of the collapsible menu for the table of contents to "TableOfContentsCollapsibleMenu".

Changed the name of the collapsible menu for the sidebar to "SidebarCollapsibleMenu".