Completed the search bar UI
Also made the colours for the main icon, though it will still need to be updated tomorrow.

Changed around a check in the collapsible menu for the sidebar so that typescript would be happy.

Made the "All Wrongs Reserved" text disappear when the screen is too small.

Updated all the packages used.