Disabled the aggressive tricks
The aggressive tricks in the tricks.ts file and the sidebar component are disabled until the data for the aggressive tricks is complete.

Also removed some categories from the aggressive tricks that don’t have any tricks.

Added a wrapper to wrap the theme toggler so the theme toggler can have padding without affecting the size of the theme toggler icon.

Changed the padding for the nav bar since the nav bar doesn’t need to pad for the theme toggler anymore.