Completed the backend for the search
Finally managed to get pagefind working by ignoring the pagefind import in the +page.ts file for the search page. The +page.ts file reads the user's query and loads the search results from pagefind since the +page.svelte file doesn't support top-level await, so all that is done within the async load function which is exported in the +page.ts file for the search page. Also managed to get svelte-check to be useful by setting checkJs in the tsconfig.json to false, otherwise the svelte-check command will just print out impossibly long error messages which makes it useless. All the files used are .ts files, so removing the .js files from svelte-check is just fine.

To get pagefind to work, the src/routes/+layout.ts file was added to get Sveltekit to pre-render the entire website, unless explicitly told to do otherwise.

The +page.ts file for the search page is excluded from pre-rendering as it is dynamically generated based on the user's query.

Also added a types.ts file to the src/lib directory which contains all the types used in the website. It now currently contains all the types used by the pagefind library.

Added a pagefind.json file so the configuration of the source directory (.svelte-kit) and the bundle directory (where the pagefind.js bundle is located) is easier to maintain. It is preferable to editing the pagefind command in the package.json file.

Moved the pagefind bundle to src/lib/.pagefind as Sveltekit has a $lib alias for the src/lib directory which makes importing pagefind far easier. With this change, the src/lib/.pagefind directory has to be ignored by git as it is generated after every build and hence shouldn't be tracked by git, so the src/lib/.pagefind directory has been added to the .gitignore file.

Removed the init:search script in the package.json file after realising that you can use the command npm run postbuild to run the postbuild script. The postbuild script just runs pagefind now, as all the configuration for pagefind is done in the pagefind.json file.

Completed the search bar component and the handleSearchSubmit function and removed the handleSearch function which was supposed to generate search results as it was too difficult to do the dropdown for the search results. Now the search bar just directs the user to the search page where all the results are displayed.

Added the search page in the src/routes/search folder.

Also added a search result component for the search page so that displaying the results would be easier.

Removed the console.log in the table of contents component.

Added a paragraph to the main page to test the mark element.

Completed the colours for the mark element and also set the anchor and mark element's opacity to 1 so that the elements won't be affected by the text CSS class, which has an opacity of 0.87 in dark mode.

Updated the packages used.