Completed the landing page and search page
Added a page left-right margin custom CSS property so I can easily set the margin on all pages.

Updated the main icon to use the new one with added sections. The main icon is still pending one change for the light theme and other than that, the main icon for the light theme is complete.

Changed the hamburger width custom CSS property as the hamburger menu is now not wider than the bar width anymore for some reason. With this change, the size of the theme toggler has also been changed to remain at 30px.

Also changed the left-right padding on the entire nav bar so that there's no need to put the padding on the hamburger menu and the theme toggler.

Also made the gap between the website icon and the website name bigger, to 10px.

Also removed the padding on the theme toggler wrapper.

Extracted the JavaScript object representing the pages in the website into a file called constants.ts and imported that object into the sidebar.

Extracted the function to make a URL-friendly string into a file called utils.ts, which will store all the utility functions used in the website, and imported the function into the sidebar collapsible menu component and the table of contents collapsible menu component.

Completed the search result component that is used on the search page.

Renamed the wrapper class in the root layout file to page-wrapper.

Edited the root layout file so that it is easier to make the content on the page centred, specifically by using display flex on the page container and the page wrapper and setting flex to 1 on the page wrapper.

Added a custom error page that would show the error and redirect the user back to the landing page after about 1 second.