Fixed the flashing when the site is refreshed
Also made the class for the theme to be placed on the HTML element instead of the document body.

Labelled the text in the error page with the text class.

Removed the flex-wrap from the page container and the page wrapper as it is not needed for the elements to be centred.

Centred the links in the link grid on the landing page.

Changed the favicon to a black-and-white version of the website icon. For some reason, it doesn't work with iOS devices using Safari as their browser. The favicon displayed on these devices is still the original Sveltekit icon.