Added a scroll up button to the layout
Created the scroll-up button component.

Added a div to mark the top of the page and make it invisible.

Moved the custom CSS property for the animation timing to the root selector instead of being in the nav bar component as the scroll-up button component is not within the nav bar.

Updated the main icon so the toe cap can be coloured.

Made the error page not display the error message if there is no error message.

Removed the colour in the styles for the telegram-bot-wrapper class.