Fixed the collapsible menus in the sidebar
Now the icons to open the collapsible menus are much easier to use on mobile as the selectable area is the entire area after the words, which makes it much easier to tap.

The position of the drop-down icons has also been shifted to the end of each header, which makes them all line up nicely compared to previously where the font would make the icons not line up as the font isn't monospace.

Also made the table of contents title change colour on hover to make it obvious to the user that the entire width of the table of contents header is clickable to open the table of contents.

Also changed the margins and the paddings on the sidebar to make it line up with the nav bar better as well as to make it look better.