Changed the HTML tags for better semantics
For the main page, placed the heading and the description on how to use the site in a header tag. Also wrapped the grid of links in a main tag instead of a div as that is where the content is.

For the terminology page, used a section element instead of a div to wrap a category.