Moved the dev contact to the start message
Moved the part about contacting the developer to the start message instead of the help message as the help message is just way too long and it's quite unlikely that people would read the first part.

Also added a part telling the user about the website in the start message.