Imported all of the aggressive tricks to the site
The pages containing the aggressive tricks have imported all of the required data, but they don't display anything as the data for the aggressive tricks isn't complete. This is so that the pagefind library doesn't index them as they have no content and users don't find the page. The aggressive trick pages are also excluded from the sidebar which makes it very unlikely for users to stumble upon the aggressive trick pages.