Renamed rinks to skating rinks
The bot will now accept skating rinks with an underscore as a valid command and the help command will also display the skating rinks command instead of the rinks command, but the rinks command will still work.

Also renamed skateparks to skate parks, so that the bot will be consistent with the website. The recommended commands for the bot in the help command have also been updated and the bot now also accepts the skate parks command with the underscore.