The boldFirstLine function now smartly strips HTML
Added the unstripValidHtmlTags function to convert valid HTML tags that have been stripped back into regular HTML tags. This function is used in the boldFirstLine function so that hyperlinks in the training message aren't lost because they are being stripped.

Added a dictionary called charToHtmlEntity containing all of the characters to convert to HTML entities. This dictionary is used to replace all the characters with their respective HTML entities.

The stripHtml function now uses the charToHtmlEntity dictionary instead of calling another function which uses a switch statement to determine what to replace.

Created a reverseDict function so that the charToHtmlEntity dictionary can be easily reversed, which is used in the unstripValidHtmlTags function.

Removed the stripHtmlReplacer function as it is no longer needed.