Moved 2 functions to the utils.ts file
The convertPathToUrl function has been renamed to "convertFilePathToUrl" and the getFilename function has been renamed to "getFilenameFromFilePath" to be more specific since the functions are moved to the utils.ts file.

Moved the convertFilePathToUrl and the getFilenameFromFilePath function from the ListsPage component to the utils.ts file as the two functions are used on other pages as well.