Imported the places enum into the discount info file
This way, the discount information file depends on the places enum in the places.ts file and will error when the name of the JSON files has been changed.

Also exported the function to load the places JSON file, which is imported into the discount info file to load the places needed for the discount information.