Made the terminology page more like the others
This way, the terminology page is easier to navigate thanks to the table of contents component generating a link for every single term, which makes it much easier to find the term that one is looking for.

It also makes the site more consistent as most of the pages look a certain way while the terminology page sticks out like a sore thumb.