Commit the pagefind folder
This is so that the deployment on Vercel (somehow vite just refuses to build there) will be able to find the pagefind.js file and not error out when it’s built. The postbuild command should run pagefind again which should override the files inside the pagefind folder to create the updated index.