Added a few functions to the utilities module
Added the function isProduction to check if the Telegram bot is being run on Vercel.

Also added a function called getJsonRoot to get the default root file path for the JSON files that depends on whether or not the bot is being run on Vercel.

The loadJsonData function now uses the getJsonRoot function to get the data folder to load the JSON files.

Also edited the loadPdfFile function to have a default folder of /static instead of /src/lib as that is where the PDF files are located.