Created a function to remove the bot’s username
Applied this function to all commands that take in user input to generate a reply. This is because sending a command from the command menu in a group automatically adds the bot’s username behind the command, which needs to be removed, otherwise the bot will end up generating a wrong reply.

Also changed the name of the qrCodeHandler function to just handler, as the module name is in front of the function when it is called in the index.ts file of the telegram bot. Hence, there is no need to repeat the word “qrCode”.

Also edited some inaccurate comments in the command handlers.