Swapped the 2nd and 3rd places
Dropped the ranking of the Nexus Phantom due to the screws on them becoming loose after 2 or 3 months, which is an issue you pretty much never need to worry about on any other skate. They also have a riveted cuff which makes it difficult to replace the cuff when it breaks so their ranking has dropped. The reason for the Nexus Phantom recommendation has been updated to reflect all these issues.

The Flying Eagle F5S Eclipse takes the Nexus Phantom's place at number 2.