Removed the unstripValidHtml function
The unstripValidHtml function is useless as regenerating the poll message will not preserve the message entities like hyperlinks and italics, which will turn the poll message back into a regular one without any message entities. It is only useful when no one votes on the poll, so it's effectively useless.