Added poll types
The poll type is a zero-width string that the bot will use to determine
what kind of poll message it is.

It is used to determine which callback query function to run when the
rental message function is implemented, since the callback query for it
is different from the regular poll message function.

Also fixed up the formatting.