Rewrote the createPollPortion function
Now the createPollPortion function should be able to support all the
features required in the rental message command, like preserving the
number of lines in the poll portion and tagging a name with a symbol.

Created a bunch of functions to support the createPollPortion function.

Also changed getPollOptionData to getPollOptionSegment as most of the
data is now handled by the createPollPortion function.

Also made the string for the poll type to appear at the end of the input
instead of in the middle of the 2 new lines so that it wouldn't mess
with the formatting of the poll message.