Added a component to load the pagefind highlighter
This component wraps all the pages by taking in the slot in the
layout.svelte file so that every time the page changes, the component
will be unmounted and remounted to get the pagefind highlighting script
to reload.

Having the pagefind highlighting script in the onMount function in the
layout means that the highlighting script will only load once, as the
layout is constant across all pages of the website, which means that
when the page has changed, the highlighting script will just no longer
work since the highlighter has already run and place all the highlights
where they need to be at and won't run again.

Also placed a data-pagefind-body attribute on the component to load the
highlighting script so that only the page contents get highlighted
instead of the layout elements as well. It shouldn't affect the search
since the content of the layout isn't included in the search any way.

Also updated the packages used.