Fixed the logic for the preserve lines option
This is in the regeneratePollPortion function. It now no longer depends
on the number of matches as that can continually increase as people add
their names. Even if people take their name off the poll, the number of
matches wouldn't change as it would detect a blank line. Hence, it now
depends on the maximum number of entries only and will not preserve the
lines if there is no maximum. It also will preserve the line if there
are no names on the poll and there's a numbering style in use, to
preserve the numbering style.