Made the delete messages function more robust
Also moved the callNextStep function to the bot utils and renamed it to
callStep, as it now can call not only the next step, but also the
previous step or a specific step that is given.

Also made the create poll message scene more customisable by passing a
list of prompts for it to prompt the user with.