Removed the regex for the poll message
The regex is really only used when the command is available in inline
mode, and since the poll message doesn't work in inline mode, there's no
need to have a regex for the poll message command.

Hence, all replacements that previously used the poll message regex have
been migrated to use the removeBotUsernameAndCommand function instead.