Updated the createNumberOfPeoplePortion function
The createNumberOfPeoplePortion function now supports substituting the
maximum number of people into the format string.

As such, the createNumberOfPeoplePortion function is now in charge of the show
remaining option instead of the functions that call it, since the
maximum number of entries have to be passed to the function anyway to
substitute in the maximum number of people.