The next function doesn't call the next step
Also made the message text to be empty in the callStep function so the
bot doesn't process the user's message once more.

This caused an infinite loop when going from the fourth step of the
create poll message scene back to the third step of the create poll
message scene, where the answer to the additional option function was
being passed as a poll option.