Updated the answerIfGlobalLimitIsHit function
The number of entries should be less than or equal to the limit to
return false as the person should only be stopped when they exceed
the limit.

Also, the reformed poll message is now taken instead of the original
poll message as there is not enough information at that point to
determine whether the entry is being added or removed from the poll.
Only after the poll message is reformed do we know that the person's
name was added or removed.