Updated the sizes for NTU
Also made the answerRentalMessageCbQuery take an optional
tagFallbackFunc which is called when the user isn't tagged, which can
happen when using a regular rental message function in a group with a
custom rental message function and the user is being tagged incorrectly.
This function makes it possible to call the default callback handler for
the message to correctly handle the tag.