Delete everything related to the tagFallbackFunc
There's no need for it at all, since the reformed poll message screws
the message up and causes the answerIfGlobalLimitIsHit function to find
more than one match for the rental message, when it really should have
only one.

Now, the NTU rental message callback handler fallsback to the default
rental message callback handler when the poll options in the poll
message isn't the same as the list of sizes. This should be good enough
because I don't think anyone would create a rental message with the exact
same poll options as the list of sizes in the NTU chat group.