Made the getPollOptions function work properly
I have no idea why I only got the first button in a row to get the poll
options. I guess there used to be no need for the feature when it was
just the regular poll message. But now with the rental message having
two buttons in a row, the getPollOptions function should iterate over
all of the buttons to get the poll options and not just the first button
on each row.