Created the default numbering style message
This is so that the list of numbering styles would be sent together with
the message instead of just relying on the reply keyboard to appear,
which doesn't appear in groups with topics.

Also, copying and pasting from the list in the message sent also makes
sure the message sent by the user is sent in the same topic, instead of
the reply keyboard sending the message to the general topic.