Removed all logging
Also, the createNumberOfPeoplePortion function shows infinity when there
is no maximum number of entries now and will substitute the maximum
number of entries regardless of whether the number of entries is present
or not.

This is to deal with the case where the person has created a rental
message but forgot to enter the number of rentals for a rental option
and used the /done command to generate the rental message.

Removed the double quotes from the numberingStyles in the strFormat in
the second step of the create poll message scene.