Remove the bot's username with the command
The removing the bot's username was only relevant when people used
Telegram's autocomplete in groups and had the bot's username tagged
behind the command.

This should remove the bot's username only after the command. All the
regex to validate the various commands have also been updated to remove
the bot's username after the command.