Make the description for the Nexus Phantom less positive
The Nexus Phantom's quality is better than the Oxelo MF500, but it's
not as much as the text used to make it seem.

Also added a reason for why the screws keep on loosening, which is the
lack of Loctite on the screws.

Still, the skate is a great buy and for people on a tight budget.