Fix TaoBao link
TaoBao links are just so god damn annoying. They are almost guaranteed to break after a while if you get your links from the TaoBao app or from a search engine like Google. They also have so many different link formats that it’s impossible to know whether a link format will not break after a while or not. But, from experience, all of the link formats that can be found or retrieved organically will break after a while.

Hopefully this link format doesn’t break, since it’s apparently used by the mainland China desktop version of the website.

The link format is:
https://item.taobao.com/item.htm?id={some_product_id}